Move @cached decorator out into synapse.storage._base; add minimal docs
